Move all annotations into package sorald.annotations
