+++ src/main/java/sorald/annotations/IncompleteProcessor.java
@@ -1,4 +1,4 @@
-package sorald.sonar;
+package sorald.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
