+++ src/main/java/sorald/annotations/ProcessorAnnotation.java
@@ -1,4 +1,4 @@
-package sorald;
+package sorald.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
