+++ src/main/java/sorald/processor/IteratorNextExceptionProcessor.java
@@ -2,7 +2,7 @@ package sorald.processor;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
-import sorald.ProcessorAnnotation;
+import sorald.annotations.ProcessorAnnotation;
 import spoon.reflect.code.CtBlock;
 import spoon.reflect.code.CtCodeSnippetExpression;
 import spoon.reflect.code.CtExpression;
