+++ src/main/java/sorald/processor/MathOnFloatProcessor.java
@@ -2,7 +2,7 @@ package sorald.processor;
 
 import java.util.List;
 import sorald.Constants;
-import sorald.ProcessorAnnotation;
+import sorald.annotations.ProcessorAnnotation;
 import spoon.reflect.code.BinaryOperatorKind;
 import spoon.reflect.code.CtBinaryOperator;
 import spoon.reflect.code.CtCodeSnippetExpression;
