+++ src/main/java/sorald/processor/PublicStaticFieldShouldBeFinalProcessor.java
@@ -1,7 +1,7 @@
 package sorald.processor;
 
-import sorald.ProcessorAnnotation;
-import sorald.sonar.IncompleteProcessor;
+import sorald.annotations.ProcessorAnnotation;
+import sorald.annotations.IncompleteProcessor;
 import spoon.reflect.declaration.CtField;
 import spoon.reflect.declaration.ModifierKind;
