+++ src/main/java/sorald/processor/SynchronizationOnGetClassProcessor.java
@@ -1,7 +1,7 @@
 package sorald.processor;
 
 import java.util.Set;
-import sorald.ProcessorAnnotation;
+import sorald.annotations.ProcessorAnnotation;
 import spoon.reflect.code.CtExpression;
 import spoon.reflect.code.CtFieldAccess;
 import spoon.reflect.code.CtInvocation;
