+++ src/main/java/sorald/processor/XxeProcessingProcessor.java
@@ -10,8 +10,8 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilderFactory;
-import sorald.ProcessorAnnotation;
-import sorald.sonar.IncompleteProcessor;
+import sorald.annotations.ProcessorAnnotation;
+import sorald.annotations.IncompleteProcessor;
 import spoon.reflect.code.*;
 import spoon.reflect.declaration.CtCompilationUnit;
 import spoon.reflect.declaration.CtImport;
